DyslexiaBackend/requirement.txt
Django
djangorestframework
django-filter
openai
django-cors-headers